pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI Boilerplate"
authors = [{name = "zhanggc", email = "zhanggc2018@qq.com"}]
dependencies = [
    "alembic>=1.9.4",
    "SQLAlchemy>=2.0.4",
    "PyJWT>=2.6.0",
    "uvicorn>=0.37.0",
    "fastapi>=0.119.0",
    "celery>=5.5.3",
    "gunicorn>=23.0.0",
    "fastapi-event>=0.1.3",
    "ujson>=5.7.0",
    "click>=8.1.3",
    "greenlet>=3.0.0",
    "redis>=4.5.4",
    "pydantic>=2.10.0,<2.12.0",
    "pydantic-settings>=2.11.0",
    "isort>=7.0.0",
    "black>=25.9.0",
    "pylint>=4.0.0",
    "asyncpg>=0.27.0",
    "pytest>=8.4.2",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "email-validator>=2.3.0",
    "python-jose>=3.5.0",
    "cryptography>=46.0.2",
    "passlib>=1.7.4",
    "bcrypt>=5.0.0",
    "faker>=37.11.0",
    "pytest-asyncio>=1.2.0",
    "behave>=1.3.3",
    "loguru>=0.7.3",
    "asgi-correlation-id>=4.3.4",
    "fastapi-limiter>=0.1.6",
    "fastapi-pagination>=0.14.3",
    "openai>=2.3.0",
    "requests>=2.32.5",
    "sync-to-async>=0.1.1",
    "user-agents>=2.2.0",
    "asgiref>=3.10.0",
    "starlette-context>=0.4.0",
    "pydantic-core>=2.27.0,<2.41.0",
    "orjson>=3.11.4",
]
requires-python = ">=3.12"
readme = "README.md"

[project.optional-dependencies]
dev = [
    "behave>=1.3.3"
]

[tool.uv]
python-install-mirror = "https://registry.npmmirror.com/-/binary/python-build-standalone/"

[[tool.uv.index]]
name = "tuna"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true


# Pyright 配置
[tool.pyright]
# 关闭类型检查
typeCheckingMode = "off"

# Ruff 常用检验规则
[tool.ruff]
# 行长度限制
line-length = 120
# 目标 Python 版本
target-version = "py312"
# 保存时自动修复
fix = true
# 自动修复不安全的修复
unsafe-fixes = true

# 格式化设置
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.ruff.lint]
# 启用核心检查规则
select = [
    "F",     # Pyflakes - 包含未定义名称检查
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "B",     # flake8-bugbear
    "N",     # pep8-naming
]

# 可以单独启用特定规则
extend-select = [
    "F821",  # 未定义名称
    "F822",  # 方法或属性不存在
]

# 设置严重等级
extend-ignore = []

# 每个文件的特定规则
[tool.ruff.lint.per-file-ignores]
"migrations/env.py" = ["E402"]  # 允许migrations文件中的模块级导入不在文件顶部
"app/api/deps/*" = ["B008"]  # 允许FastAPI依赖注入中的函数调用
"app/api/v1/*" = ["B008"]  # 允许API路由中的FastAPI依赖注入
